revis terkait dp, bersih file, status pembayaran + diskon

+++ database/migrations/2025_10_03_133231_create_cafe_products_table.php
@@ -10,8 +10,8 @@ return new class extends Migration {
             $table->id();
             $table->string('nama');
             $table->string('kategori')->nullable();
-            $table->string('satuan',30)->default('porsi');
-            $table->decimal('harga_jual',12,2)->default(0);
+            $table->string('satuan', 30)->default('porsi');
+            $table->unsignedBigInteger('harga_jual')->default(0);
             $table->integer('stok')->default(0);
             $table->integer('minimal_stok')->default(0);
             $table->boolean('aktif')->default(true);
